Conflicts:
	src/Nio/EngineMgr.cpp

+++ src/Params/PADnoteParameters.cpp
@@ -458,7 +458,8 @@ void PADnoteParameters::generatespectrum_bandwidthMode(float *spectrum,
         if(ibw > profilesize) { //if the bandwidth is larger than the profilesize
             float rap   = sqrt((float)profilesize / (float)ibw);
             int   cfreq =
-                (int) (realfreq / (synth->samplerate_f * 0.5f) * size) - ibw / 2;
+                (int) (realfreq
+                       / (synth->samplerate_f * 0.5f) * size) - ibw / 2;
             for(int i = 0; i < ibw; ++i) {
                 int src    = (int)(i * rap * rap);
                 int spfreq = i + cfreq;
@@ -476,7 +477,7 @@ void PADnoteParameters::generatespectrum_bandwidthMode(float *spectrum,
                 float idfreq = i / (float)profilesize - 0.5f;
                 idfreq *= ibw;
                 int   spfreq  = (int) (idfreq + ibasefreq);
-                float fspfreq = fmod((double)idfreq + ibasefreq, 1.0f);
+                float fspfreq = fmodf((float)idfreq + ibasefreq, 1.0f);
                 if(spfreq <= 0)
                     continue;
                 if(spfreq >= size - 1)
